+++ src/app/app.component.html
@@ -1,6 +1,6 @@
 <div class="container">
   <div class="row">
-    <div class="col-12 spinner">
+    <div class="col-12">
     <app-loading-spinner></app-loading-spinner>
   </div>  
     <div class="col-12 ">
