+++ src/app/app.routes.ts
@@ -1,20 +1,18 @@
 import { Routes } from '@angular/router';
 import { HomeComponent } from './home/home.component';
-import { CountryResolver } from './resolver/country.resolver';
 import { CountryDataResolver } from './resolver/country-data.resolver';
 
 export const routes: Routes = [
   { path: '', redirectTo: '/country', pathMatch: 'full' }, //default route
-
   {
     path: 'country',
     component: HomeComponent,
     resolve: { countries: CountryDataResolver },
-  }, // Country listing route
+  },
   {
     path: 'country/:name',
     component: HomeComponent,
     resolve: { countries: CountryDataResolver },
-  }, // Country name route when open details dialog
+  }, 
   { path: '**', redirectTo: '', pathMatch: 'full' },
 ];
