+++ src/app/resolver/country-data.resolver.ts
@@ -5,13 +5,12 @@ import { Store } from '@ngrx/store';
 import { Country } from '../models/country.model';
 import { CountryService } from '../services/country.service';
 import { loadCountries } from '../store/country.action';
-import { catchError, tap } from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root',
 })
 export class CountryDataResolver implements Resolve<Country[]> {
-  constructor(private countryService: CountryService, private store: Store) {}
+  constructor(private store: Store) {}
 
   resolve(): any {
     this.store.dispatch(loadCountries());
