+++ src/app/shared/loading-spinner/loading-spinner.component.html
@@ -1,1 +1,1 @@
-<mat-spinner *ngIf="(showLoading$ | async)" [value]="20" ></mat-spinner>
+<mat-spinner *ngIf="(showLoading$ | async)" [value]="20" class="spinner"></mat-spinner>
