+++ src/app/store/country.selector.ts
@@ -1,5 +1,5 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
-import { CountryState } from './country.reducer';
+import { CountryState } from './country.state';
 
 export const selectCountryState =
   createFeatureSelector<CountryState>('countries');
@@ -9,13 +9,11 @@ export const selectAllCountries = createSelector(
   (state: CountryState) => state.countries
 );
 
-// Selector to get a selected country by name
 export const selectCountryByName = (countryName: string) =>
   createSelector(selectAllCountries, (countries) =>
     countries.find((country) => country.name.common === countryName)
   );
 
-// Selector to filter countries based on a search term
 export const selectFilteredCountries = (
   searchTerm: string,
   searchBy: 'name' | 'capital'
